Coverage

+++ photo-server/src/main/java/org/stapledon/security/entities/UserInfo.java
@@ -24,23 +24,29 @@ import java.util.Set;
 public class UserInfo {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer userId;
+    private Long userId;
+
     @NotBlank
     @Size(max = 50)
     @Email
     private String email;
+
     @NotBlank
     @Size(max = 120)
     private String password;
+
     @NotBlank
     @Size(max = 20)
     private String username;
+
     @NotBlank
     @Size(max = 20)
     private String firstName;
+
     @NotBlank
     @Size(max = 20)
     private String lastName;
+
     @ManyToMany(fetch = FetchType.LAZY)
     @JoinTable(name = "user_role",
             joinColumns = @JoinColumn(name = "user_id"),
